refactor(settings): move Update to About Exodus, add GitHub/X/License links
Move UpdatePanel and Auto Update toggle from General settings to the About
Exodus page, and enrich it with GitHub, X (Twitter), and License links.
