+++ src/renderer/components/settings/settings-form/system-info.tsx
@@ -1,9 +1,44 @@
+import { UseFormReturnType } from '@shared/schemas/settings-schema'
+import { ExternalLinkIcon } from 'lucide-react'
+import { useEffect } from 'react'
+
+import { Switch } from '@/components/ui/switch'
+import { useUpdater } from '@/hooks/use-updater'
+import { updaterSetAutoDownload } from '@/lib/ipc'
+
 import { version } from '../../../../../package.json'
 import { SettingsRow, SettingsSection } from '../settings-row'
+import { UpdatePanel } from './update-panel'
 
-export function SystemInfo() {
+function ExternalLink({
+  href,
+  children
+}: {
+  href: string
+  children: React.ReactNode
+}) {
+  return (
+    <a
+      href={href}
+      target="_blank"
+      rel="noopener noreferrer"
+      className="text-ring inline-flex items-center gap-1 text-sm hover:underline"
+    >
+      {children}
+      <ExternalLinkIcon size={12} />
+    </a>
+  )
+}
+
+export function SystemInfo({ form }: { form: UseFormReturnType }) {
   const { versions } = window.electron.process
   const { os } = window.api
+  const { payload } = useUpdater()
+  const autoUpdate = form.watch('autoUpdate') ?? true
+
+  useEffect(() => {
+    updaterSetAutoDownload(autoUpdate)
+  }, [autoUpdate])
 
   return (
     <SettingsSection>
@@ -25,6 +60,32 @@ export function SystemInfo() {
       <SettingsRow label="OS">
         <span className="text-ring text-sm">{os}</span>
       </SettingsRow>
+
+      <SettingsRow label="GitHub">
+        <ExternalLink href="https://github.com/exodus-ai-org/exodus">
+          exodus-ai-org/exodus
+        </ExternalLink>
+      </SettingsRow>
+      <SettingsRow label="X (Twitter)">
+        <ExternalLink href="https://x.com/ExodusAIOrg">
+          @ExodusAIOrg
+        </ExternalLink>
+      </SettingsRow>
+      <SettingsRow label="License">
+        <span className="text-ring text-sm">MIT</span>
+      </SettingsRow>
+
+      <SettingsRow
+        label="Auto Update"
+        description="Automatically download and install updates when available"
+      >
+        <Switch
+          checked={autoUpdate}
+          onCheckedChange={(checked) => form.setValue('autoUpdate', checked)}
+        />
+      </SettingsRow>
+
+      <UpdatePanel payload={payload} autoUpdate={autoUpdate} />
     </SettingsSection>
   )
 }
